Requiring Translator Functions.
1. Changed Skill Manager from service to Class
2. Relocated Skill Manager to Shared folder
3. Accessing the Skill Translator Object of the skill Manager Class in Base Component
