+++ client/app/step-builder/shared/skill-manager.service.ts
@@ -1,0 +1,13 @@
+class SkillManagerClass {
+  skillTranslator: Object;
+  constructor() { 
+    this.skillTranslator = this.getSkillTranslator();
+  }
+
+  getSkillTranslator(){
+        this.skillTranslator = require("../../../../server/libs/skills/xl/moveCellContent/moveCellContent.js");
+        return this.skillTranslator;
+  }
+}
+
+export let skillManager = new SkillManagerClass();
